Move portal link resource
