+++ workos/types/portal/__init__.py
@@ -1,0 +1,2 @@
+from .portal_link_intent import *
+from .portal_link import *
